fix: resolve AWE_OS kernel & services compilation and test suite failures
- Updated toolchain setup in rust-toolchain.toml to nightly for no_std workspace support.
- Fixed const trait operations and type derives across capability and completion modules.
- Added missing workspace dependencies to awe-product-core-tests.
- Corrected 32-bit BAR mask expansion and VirtIO legacy/modern PCI device ID checks in awe-driverd.
- Updated supervisor restart state transitions and ARP test byte pattern fixtures.
- Cleaned up clippy warnings in process, net, and policy modules.
- Ensured build target directory is untracked in .gitignore.

+++ services/update/src/transaction.rs
@@ -1,5 +1,3 @@
-#![no_std]
-
 use crate::{Slot, SlotState, UpdateError, UpdateManager, UpdateManifest};
 
 /// Crash-safe update transaction metadata. A persistent backend can journal
@@ -18,7 +16,7 @@ impl Transaction {
         target: Slot,
         manifest: UpdateManifest,
     ) -> Result<Self, UpdateError> {
-        if target == manager.active() {
+        if target as u8 == manager.active() as u8 {
             return Err(UpdateError::InvalidTransition);
         }
         if manifest.generation < manager.generation() {
